health: #100 rename & remove stuff in utils
