+++ src/model/instance.ts
@@ -1,12 +1,10 @@
-import { NullableAttribute } from "..";
-import { Class, Primitive } from "../utils";
-import { ArrayAttribute, Discriminant, Property, RequiredAttribute } from "./property";
-
-type Unbox<T> = T extends any[] ? T[number] : T;
+import { Class, Primitive, Unbox } from "../utils";
+import { ArrayAttribute, Discriminant, Property, RequiredAttribute, NullableAttribute } from "./property";
 
 type InstancePropertyValueType<V> = V extends Primitive ? ReturnType<V> : V extends Discriminant ? V : V extends Class ? Instance<InstanceType<V>> : never;
 type BoxIfArray<P, V> = P extends ArrayAttribute ? V[] : V;
-type AlsoNullIfNullable<P, V> = P extends NullableAttribute ? V | null : V;
+type NullIfNullable<P, V> = P extends NullableAttribute ? V | null : V;
+
 export type RequiredPropertyKeys<T> = Exclude<{ [K in keyof T]: T[K] extends RequiredAttribute ? K : never }[keyof T], undefined>;
 
 type InstanceDefault<T> = {
@@ -15,13 +13,11 @@ type InstanceDefault<T> = {
     // more convenient to simply supply a plain old javascript object for a function that wants an expansion.
     // an incredibly relief is that we can still use mapped keyof stuff like in "InstanceRequired<T>";
     // looks like TypeScript can provide intellisense if at least one of the types has "keyof T"
-    [K in keyof T]?: T[K] extends Property<infer V> ? AlsoNullIfNullable<T[K], BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>>> : T[K];
-    // [K in keyof T]?: T[K] extends Property<infer V> ? BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>> : T[K];
+    [K in keyof T]?: T[K] extends Property<infer V> ? NullIfNullable<T[K], BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>>> : T[K];
 };
 
 type InstanceRequired<T> = {
-    [K in RequiredPropertyKeys<T>]: T[K] extends Property<infer V> ? AlsoNullIfNullable<T[K], BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>>> : T[K];
-    // [K in RequiredPropertyKeys<T>]: T[K] extends Property<infer V> ? BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>> : T[K];
+    [K in RequiredPropertyKeys<T>]: T[K] extends Property<infer V> ? NullIfNullable<T[K], BoxIfArray<T[K], InstancePropertyValueType<Unbox<V>>>> : T[K];
 };
 
 export type Instance<T> = InstanceDefault<T> & InstanceRequired<T>;
