+++ src/utils/get-instance-class.fn.ts
@@ -1,4 +1,4 @@
-import { Class } from "./lang";
+import { Class } from "./types";
 
 export function getInstanceClass<T>(instance: T): Class<T> {
     const ctor = (instance as any).constructor;
