+++ src/utils/index.ts
@@ -1,9 +1,9 @@
-export * from "./get-instance-class";
+export * from "./get-instance-class.fn";
 export * from "./is-defined.fn";
 export * from "./is-instance-of.fn";
 export * from "./is-not-null.fn";
-export * from "./is-primitive";
+export * from "./is-primitive.fn";
 export * from "./is-string.fn";
-export * from "./lang";
+export * from "./types";
 export * from "./permutate-entries.fn";
-export * from "./subtract-sets";
+export * from "./subtract-sets.fn";
