+++ src/utils/is-instance-of.fn.ts
@@ -1,4 +1,4 @@
-import { Class } from "./lang";
+import { Class } from "./types";
 
 export function isInstanceOf<T>(cls: Class<T>): (value: unknown) => value is T {
     return (value => value instanceof cls) as (value: unknown) => value is T;
