+++ src/utils/is-primitive.fn.ts
@@ -1,4 +1,4 @@
-import { Primitive } from "./lang/primitive";
+import { Primitive } from "./types/primitive";
 
 export function isPrimitive(x?: any): x is Primitive {
     return x === Boolean || x === Number || x === String;
