--- src/utils/lang/index.ts
@@ -1,9 +1,0 @@
-export * from "./box";
-export * from "./class";
-export * from "./fields";
-export * from "./json";
-export * from "./merge-union";
-export * from "./partial-fields";
-export * from "./primitive";
-export * from "./replace";
-export * from "./unbox";
