--- src/utils/lang/partial-fields.ts
@@ -1,3 +1,0 @@
-import { Fields } from "./fields";
-
-export type PartialFields<T> = Partial<Pick<T, Fields<T>>>;
