--- src/utils/lang/readme.md
@@ -1,1 +1,0 @@
-this folder should only contain TypeScript types
