--- src/utils/lang/replace.ts
@@ -1,1 +1,0 @@
-export type Replace<T, K extends keyof T, V> = Omit<T, K> & Record<K, V>;
