+++ src/utils/types/class.ts
