+++ src/utils/types/unbox.ts
@@ -1,0 +1,1 @@
+export type Unbox<T> = T extends any[] ? T[number] : T;
